Fix a number of compiler warnings.

Related to #917

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@11149 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/pbespgsolve/Graph_impl.h
@@ -62,7 +62,7 @@ void StaticGraph::make_subgraph( const StaticGraph &graph,
                                  ForwardIterator vertices_end )
 {
     // FIXME: determine which cut-off value works best:
-    if (std::distance(vertices_begin, vertices_end) < graph.V()/static_cast<verti>(3))
+    if (static_cast<size_t>(std::distance(vertices_begin, vertices_end)) < graph.V()/static_cast<verti>(3))
     {
         HASH_MAP(verti, verti) map;
         return make_subgraph(graph, vertices_begin, vertices_end, map);
